security/requirements.txt
pytesseract
pillow
opencv-python
python-dotenv
tabulate
python-docx
docx2pdf
pymupdf
